+++ src/OpenLoco/src/Ui/Windows/Construction/Construction.h
@@ -179,7 +179,7 @@ namespace OpenLoco::Ui::Windows::Construction
         void previousTab(Window* self);
         void nextTab(Window* self);
 
-        template<uint32_t NewCapacity, uint32_t LegacyCapacity>
+        template<size_t NewCapacity, size_t LegacyCapacity>
         void copyToLegacyList(const sfl::static_vector<uint8_t, NewCapacity>& sflType, uint8_t (&legacyList)[LegacyCapacity])
         {
             static_assert(LegacyCapacity > NewCapacity);
